feat: 코드 리뷰 반영 #4 - 패키지 명 변경 - 컨트롤러의 비즈니스 로직을 서비스 단으로 이동 - ReviewServiceTest: @ExtendWith(MockitoExtension.class)로 변경 - Mockito의 when() -> given()으로 변경
