--- src/main/java/com/example/demo/domain/review/application/dto/response/GetReviewsResponse.java
@@ -1,13 +1,0 @@
-package com.example.demo.domain.review.application.dto.response;
-
-import com.example.demo.domain.review.persistence.Review;
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-
-import java.util.List;
-
-@Getter
-@AllArgsConstructor
-public class GetReviewsResponse {
-    private List<Review> reviews;
-}
