+++ src/main/java/io/cocotrip/BeApplication.java
@@ -1,0 +1,13 @@
+package io.cocotrip;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class BeApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(BeApplication.class, args);
+	}
+
+}
