+++ src/main/java/io/cocotrip/domain/review/application/dto/request/ModifyReviewRequest.java
@@ -1,0 +1,14 @@
+package io.cocotrip.domain.review.application.dto.request;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Getter
+@AllArgsConstructor
+@NoArgsConstructor
+public class ModifyReviewRequest {
+    private int score;
+    private String content;
+    private String imgUrl;
+}
