+++ src/main/java/io/cocotrip/domain/review/application/dto/request/WriteReviewRequest.java
@@ -1,0 +1,13 @@
+package io.cocotrip.domain.review.application.dto.request;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+
+@Getter
+@AllArgsConstructor
+public class WriteReviewRequest {
+    private Long placeId;
+    private int score;
+    private String content;
+    private String imgUrl;
+}
