+++ src/main/java/io/cocotrip/domain/review/application/dto/response/GetReviewsResponse.java
@@ -1,0 +1,13 @@
+package io.cocotrip.domain.review.application.dto.response;
+
+import io.cocotrip.domain.review.persistence.Review;
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+
+import java.util.List;
+
+@Getter
+@AllArgsConstructor
+public class GetReviewsResponse {
+    private List<Review> reviews;
+}
