+++ src/main/java/io/cocotrip/global/common/ApiResponse.java
@@ -1,0 +1,25 @@
+package io.cocotrip.global.common;
+
+import com.fasterxml.jackson.annotation.JsonInclude;
+import lombok.Getter;
+import lombok.RequiredArgsConstructor;
+
+@Getter
+@RequiredArgsConstructor
+@JsonInclude(JsonInclude.Include.NON_NULL)
+public class ApiResponse<T> {
+    private Integer code;
+    private String message;
+    private T result;
+
+    public ApiResponse(Integer code, String message) {
+        this.code = code;
+        this.message = message;
+    }
+
+    public ApiResponse(Integer code, String message, T result) {
+        this.code = code;
+        this.message = message;
+        this.result = result;
+    }
+}
