+++ src/test/java/io/cocotrip/BeApplicationTests.java
@@ -1,0 +1,13 @@
+package io.cocotrip;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class BeApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
